Sqlachemy 2.0

+++ src/bot/exchange/base.py
@@ -1,5 +1,6 @@
 import abc
 import re
+from decimal import Decimal
 from typing import Union
 
 from src.bot.types.margin_type import MarginType
@@ -73,7 +74,7 @@ class BaseExchange(metaclass=abc.ABCMeta):
         pass
 
     @abc.abstractmethod
-    def get_base_amount(self, pair: str, quote_amount: float):
+    def get_base_amount(self, pair: str, quote_amount: Decimal):
         pass
 
     @abc.abstractmethod
